Remove '-run' flag and always use local copy of nob.h
We can't use '-run' flag as intended because we need the command
passed to Nob_Cmd their working directory for it to operate as
expected but we can't do that. We could possibly in the future just
extend nob locally or write a custom function for this action.

Also when calling curl for the latest nob.h version just save it to
the cache directory then retrieve it from cache into the project.

Re-consider if we should be doing network request by default cause
that's kinda extra. We could just provide a '-latest' flag for when
they want to fetch the latest one.